chore: rename screenshot-related files
